Add new archive --update option to update existing items
Add a new `--update` option to the `archive` command which checks
existing items to see if they have substantively changed, and if so,
saves the updated item.

This can be combined with `--no-stopping` to update all items in the
archive.

To make this work, add a new `recursive_compare` function which knows
how to smartly compare two objects, ignoring fields which are known to
be transient and frequently updated in inconsequential ways; and how
to deal with date/datetime fields in either structured or string
format.